+++ src/jscc/revars.js
@@ -1,7 +1,6 @@
 /**
  * @module regexlist
  */
-export { JS_STRING as STRINGS } from '../util/regexes'
 
 // name=value in directives - $1:name, $2:value (including any comment)
 export const VARPAIR = /^\s*(_[0-9A-Z][_0-9A-Z]*)\s*=?(.*)/
@@ -14,3 +13,6 @@ export const EVLVARS = /(^|[^$\w\.])(_[0-9A-Z][_0-9A-Z]*)\b(?=[^$\w]|$)/g
 
 // replace varnames inside the code from $_VAR.prop to value
 export const REPVARS = /(?:(\$_[0-9A-Z][_0-9A-Z]*)(\.[\w]+)?)(?=[\W]|$)/g
+
+// matches single and double quoted strings, take care about embedded eols
+export const STRINGS = /"[^"\n\r\\]*(?:\\(?:\r\n?|[\S\s])[^"\n\r\\]*)*"|'[^'\n\r\\]*(?:\\(?:\r\n?|[\S\s])[^'\n\r\\]*)*'/g