+++ src/util/regexes.js
@@ -1,1 +1,0 @@
-/home/alberto/Proyectos/lib/perf-regexes/src/index.js